Create Binary Search Along with Time complexity
A program that implements binary search algorithm to find a use entered name in an array of 31 elements at least. You need to populate the array with names at the start of the program and then sort it before implementing binary search on it. Print the time to perform the whole operation of binary search at the end. 
	If the name exits
o	Print index values
	If the name does not exit